Refactored private unit tests to be conditionaly built from main library. Updated README
