+++ README.md
@@ -2,7 +2,4 @@
 [iSAX 2.0](http://www.cs.ucr.edu/~eamonn/iSAX_2.0.pdf "iSAX 2.0")
 
 #Release build
-    cmake DIR && make
-
-#Unit testing
-    cmake DIR -DSTS_COMPILE_UNIT_TESTS=1 && make && make test
+    cmake DIR -DCMAKE_BUILD_TYPE=Release && make
