+++ include/symtseries.h
@@ -33,18 +33,4 @@ sax_word sts_to_iSAX(double *series, size_t n_values, size_t w, unsigned int c);
  */
 double sts_mindist(sax_word a, sax_word b, size_t n, size_t w, unsigned int c);
 
-#ifdef STS_COMPILE_UNIT_TESTS
-
-char *test_get_symbol_zero();
-
-char *test_get_symbol_breaks();
-
-char *test_to_iSAX_normalization();
-
-char *test_to_iSAX_sample();
-
-char *test_to_iSAX_stationary();
-
-#endif // STS_COMPILE_UNIT_TESTS
-
 #endif
