+++ src/test/CMakeLists.txt
@@ -3,6 +3,3 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 add_library(sts_test SHARED sts_test.c)
-add_executable(test_private_sts test_private_sts.c)
-target_link_libraries(test_private_sts symtseries sts_test)
-add_test(NAME test_private_sts WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR} COMMAND test_private_sts)
