+++ src/test/sts_test.c
@@ -2,4 +2,3 @@
 
 int mu_tests_run = 0;
 char mu_err[MU_ERR_LEN] = { 0 };
-
